+++ mousecontroller.h
@@ -21,7 +21,7 @@ signals:
     void updateViewport();
     void setViewportActive(Model::ViewportType type);
     void setClickedId(int x, int y);
-    void setActivePrimitive(int id);
+    void setActivePrimitive(float id);
 
 public slots:
